[gpu.device] API for iterating/auto-choosing hardware, and using that hardware to create a device. All new code has been well-documented and implemented for Vulkan.

+++ include/tz/core/handle.hpp
@@ -6,7 +6,7 @@
 
 namespace tz
 {
-	enum class hanval : std::uint32_t{};
+	enum class hanval : std::uint64_t{};
 
 	struct nullhand_t
 	{
